backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
requests>=2.28.0
python-dotenv>=1.0.0
google-generativeai>=0.6.0
modal>=0.62.0
